requirements.txt
django==4.0
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.6.0
social-auth-core==4.3.0
social-auth-app-django==5.1.0
google-auth==2.15.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
Pillow==9.4.0